# src/main/resources/swagger/employeeAppraisals-api.yml
...
swagger: '2.0'
...
info:
  title: |
    GFT Employee Appraisals API
  description: |
    GFT Employee Appraisals API specification
  version: 1.0.0

# URL Information
host: example.gft.com
schemes:
  - https

# Security Definition
securityDefinitions:
  OAuth2:
    type: oauth2
    flow: accessCode
    authorizationUrl: http://localhost:8888/uaa/oauth/authorize
    tokenUrl: http:/localhost:8888/uaa/oauth/token
    scopes:
      employee: Employee access
      appraisal: Appraisal access
security:
  - OAuth2: [employee, appraisal]

# Endpoints
paths:
...
  '/me/appraisals':

    ###########
    # GET #####
    ###########

    get:
      tags:
      - Appraisal
      summary: |
        Provides a list of appraisals for the authenticated user.
      description: |
        Returns a list of `Appraisal` belonging to the authenticated user.
      operationId: meAppraisalsGet
      produces:
        - application/json
      responses:
        '200':
          description: |
            Successful.
          schema:
            type: array
            items:
              $ref: '#/definitions/AppraisalDTO'
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
        '404':
          description: |
            The user is authenticated, but is not yet registered in the system.

########################################################################################################################
# /me/appraisals/:id ###################################################################################################
########################################################################################################################
  '/me/appraisals/{appraisalId}':

    ###########
    # GET #####
    ###########

    get:
      tags:
      - Appraisal
      summary: |
        Provides a specific appraisal for the authenticated user.
      description: |
        Returns an `Appraisal` belonging to the authenticated user.
      operationId: meAppraisalsIdGet
      produces:
        - application/json
      parameters:
        - in: path
          name: appraisalId
          type: integer
          required: true
          description: |
            `Appraisal` Id.
      responses:
        '200':
          description: |
            Successful.
          schema:
            $ref: '#/definitions/AppraisalDTO'
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
        '404':
          description: |
            Appraisal not found in the system.

########################################################################################################################
# /me/appraisals/:id/forms #############################################################################################
########################################################################################################################
  '/me/appraisals/{appraisalId}/forms':

    ###########
    # GET #####
    ###########

    get:
      tags:
      - Appraisal
      summary: |
        Provides a list of `EvaluationForm` of a specific appraisal for the authenticated user.
      description: |
        Returns a list of `EvaluationForm` belonging to the authenticated user.
      operationId: meAppraisalsIdFormsGet
      produces:
        - application/json
      parameters:
        - in: path
          name: appraisalId
          type: integer
          required: true
          description: |
            `Appraisal` Id.
      responses:
        '200':
          description: |
            Successful.
          schema:
            type: array
            items:
              $ref: '#/definitions/EmployeeEvaluationFormDTO'
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
        '404':
          description: |
            Appraisal not found in the system.

########################################################################################################################
# /me/appraisals/:id/forms/:id #########################################################################################
########################################################################################################################
  '/me/appraisals/{appraisalId}/forms/{formId}':

    ###########
    # GET #####
    ###########

    get:
      tags:
      - Appraisal
      summary: |
        Provides a specific `EvaluationForm` of a specific appraisal for the authenticated user.
      description: |
        Returns a particular `EvaluationForm` belonging to the authenticated user.
      operationId: meAppraisalsIdFormsIdGet
      produces:
        - application/json
      parameters:
        - in: path
          name: appraisalId
          type: integer
          required: true
          description: |
            `Appraisal` Id.
        - in: path
          name: formId
          type: integer
          required: true
          description: |
            `Form` Id.
      responses:
        '200':
          description: |
            Successful.
          schema:
            $ref: '#/definitions/EvaluationFormTemplateDTO'
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
        '404':
          description: |
            Appraisal or EvaluationForm not found in the system.

########################################################################################################################
# /me/mentor ###########################################################################################################
########################################################################################################################
...
  '/employees/{employeeId}/appraisals':

    ###########
    # GET #####
    ###########

    get:
      tags:
      - Appraisal
      summary: |
        Returns the appraisals of the specified employee.
      description: |
        Returns an `Appraisal` list belonging to the specified employee
      operationId: employeesIdAppraisalsGet
      produces:
        - application/json
      parameters:
        - in: path
          name: employeeId
          required: true
          type: integer
          description: |
            `Employee` Id.
      responses:
        '200':
          description: |
            Successful.
          schema:
            type: array
            items:
              $ref: '#/definitions/AppraisalDTO'
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
        '404':
          description: |
            An `Employee` with the specified Id was not found.

########################################################################################################################
# /employees/{employeeId}/appraisals/{appraisalId} #####################################################################
########################################################################################################################
  '/employees/{employeeId}/appraisals/{appraisalId}':

    ###########
    # GET #####
    ###########

    get:
      tags:
      - Appraisal
      summary: |
        Returns a certain appraisal of the specified employee.
      description: |
        Returns a single `Appraisal` belonging to the specified employee
      operationId: employeesIdAppraisalsIdGet
      produces:
        - application/json
      parameters:
        - in: path
          name: employeeId
          required: true
          type: integer
          description: |
            `Employee` Id.
        - in: path
          name: appraisalId
          required: true
          type: integer
          description: |
            `Appraisal` Id.
      responses:
        '200':
          description: |
            Successful.
          schema:
            $ref: '#/definitions/AppraisalDTO'
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
        '404':
          description: |
            An `Employee` or `Appraisal` with the specified Id was not found.

#########################################################################################################################
# /employees/:id/appraisals/:id/forms ###################################################################################
#########################################################################################################################
  '/employees/{employeeId}/appraisals/{appraisalId}/forms':

    ###########
    # GET #####
    ###########

    get:
      tags:
      - Appraisal
      summary: |
        Provides a list of `EvaluationForm` of a specific appraisal for the specified user.
      description: |
        Returns a list of `EvaluationForm` belonging to the specified user.
      operationId: employeesIdAppraisalsIdFormsGet
      produces:
        - application/json
      parameters:
        - in: path
          name: employeeId
          type: integer
          required: true
          description: |
            `Employee` Id.
        - in: path
          name: appraisalId
          type: integer
          required: true
          description: |
            `Appraisal` Id.
      responses:
        '200':
          description: |
            Successful.
          schema:
            type: array
            items:
              $ref: '#/definitions/EmployeeEvaluationFormDTO'
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
        '404':
          description: |
            Employee or Appraisal not found in the system.
    ###########
    # PUT #####
    ###########

    put:
      tags:
      - Appraisal
      summary: |
         Registers or updates an `Employee on `EvaluationForm` of a specific appraisal for the specified user.
      description: |
       Registers or updates an `Employee on `EvaluationForm` belonging to the specified user.
      operationId: employeesIdAppraisalsIdFormsPut
      produces:
        - application/json
      parameters:
        - in: path
          name: employeeId
          type: integer
          required: true
          description: |
            `Employee` Id.
        - in: path
          name: appraisalId
          type: integer
          required: true
          description: |
            `Appraisal` Id.
        - in: body
          name: employee
          schema:
            $ref: '#/definitions/EmployeeEvaluationFormDTO'
          required: true
          description: |
           `Employee` to create or update.

      responses:
        '200':
          description: |
            Successful.
          schema:
            $ref: '#/definitions/EmployeeEvaluationFormDTO'
        '400':
          description: |
            Input was malformed.
          schema:
            $ref: '#/definitions/EmployeeEvaluationFormDTO'
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
        '404':
          description: |
            Employee, Appraisal or EvaluationForm not found in the system.

########################################################################################################################
# /employees/:id/appraisals/:id/forms/:id ##############################################################################
########################################################################################################################
  '/employees/{employeeId}/appraisals/{appraisalId}/forms/{formId}':

    ###########
    # GET #####
    ###########

    get:
      tags:
      - Appraisal
      summary: |
        Provides a specific `EvaluationForm` of a specific appraisal for the specified user.
      description: |
        Returns a particular `EvaluationForm` belonging to the specified user.
      operationId: employeesIdAppraisalsIdFormsIdGet
      produces:
        - application/json
      parameters:
        - in: path
          name: employeeId
          type: integer
          required: true
          description: |
            `Employee` Id.
        - in: path
          name: appraisalId
          type: integer
          required: true
          description: |
            `Appraisal` Id.
        - in: path
          name: formId
          type: integer
          required: true
          description: |
            `Form` Id.
      responses:
        '200':
          description: |
            Successful.
          schema:
            $ref: '#/definitions/EvaluationFormTemplateDTO'
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
        '404':
          description: |
            Employee, Appraisal or EvaluationForm not found in the system.


    ###########
    # PUT #####
    ###########

    put:
      tags:
      - Appraisal
      summary: |
         Registers or updates an `Employee on `EvaluationForm` of a specific appraisal for the specified user.
      description: |
       Registers or updates an `Employee on `EvaluationForm` belonging to the specified user.
      operationId: employeesIdAppraisalsIdFormsIdPut
      produces:
        - application/json
      parameters:
        - in: path
          name: employeeId
          type: integer
          required: true
          description: |
            `Employee` Id.
        - in: path
          name: appraisalId
          type: integer
          required: true
          description: |
            `Appraisal` Id.
        - in: path
          name: formId
          type: integer
          required: true
          description: |
            `Form` Id.
        - in: body
          name: employee
          schema:
              $ref: '#/definitions/EvaluationFormTemplateDTO'
          required: true
          description: |
              `Employee` to create or update.
      responses:
        '200':
          description: |
            Successful.
          schema:
            $ref: '#/definitions/EvaluationFormTemplateDTO'
        '400':
          description: |
            Input was malformed.
          schema:
            $ref: '#/definitions/EvaluationFormTemplateDTO'
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
        '404':
          description: |
            Employee, Appraisal or EvaluationForm not found in the system.


########################################################################################################################
# /employees/:id/relationships #########################################################################################
########################################################################################################################
...
definitions:
  ApplicationRoleDTO:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      description:
        type: string
    required:
      - id
  AppraisalDTO:
    type: object
    required:
      - name
    properties:
      id:
        type: integer
      name:
        type: string
      description:
        type: string
      startDate:
        type: string
        format: 'date-time'
      endDate:
        type: string
        format: 'date-time'
  AnswerDTO:
      type: object
      properties:
        id:
          type: integer
        evaluationFormTemplateXSectionXQuestionDTO:
          $ref: '#/definitions/EvaluationFormTemplateXSectionXQuestionDTO'
        scoreValue:
          $ref: '#/definitions/ScoreValueDTO'
        comment:
          type: string
      required:
        - id
  EmployeeDTO:
    type: object
    properties:
      id:
        type: integer
      jobLevel:
        $ref: '#/definitions/JobLevelDTO'
      applicationRole:
        $ref: '#/definitions/ApplicationRoleDTO'
      email:
        type: string
      gftIdentifier:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      isAdmin:
        type: boolean
      isMentor:
        type: boolean
      isPeer:
        type: boolean
  EmployeeEvaluationFormDTO:
    type: object
    properties:
      id:
        type: integer
      employee:
        $ref: '#/definitions/EmployeeDTO'
      filledByEmployee:
        $ref: '#/definitions/EmployeeDTO'
      mentor:
        $ref: '#/definitions/EmployeeDTO'
      createDate:
        type: string
        format: 'date-time'
      submitDate:
        type: string
        format: 'date-time'
      comments:
        type: string
      evaluationFormId:
        type: integer
      answers:
        type: array
        items:
         $ref: '#/definitions/AnswerDTO'
  EvaluationFormTemplateXSectionXQuestionDTO:
    type: object
    properties:
      id:
        type: integer
      question:
       type: array
       items:
         $ref: '#/definitions/QuestionDTO'
    required:
      - id
...
  EvaluationFormTemplateDTO:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      description:
        type: string
      sections:
        type: array
        items:
          $ref: '#/definitions/SectionDTO'
...
  JobFamilyDTO:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      description:
        type: string
    required:
      - id
  JobLevelDTO:
    type: object
    properties:
      id:
        type: integer
      jobFamily:
        $ref: '#/definitions/JobFamilyDTO'
      name:
        type: string
      description:
        type: string
      expertise:
        type: string
    required:
      - id
...
  QuestionDTO:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      description:
        type: string
      position:
        type: integer
...
  ScoreTypeDTO:
    type: object
    properties:
      id:
        type: integer
      definition:
        type: string
      scoreValues:
        type: array
        items:
          $ref: '#/definitions/ScoreValueDTO'
  ScoreValueDTO:
    type: object
    properties:
      id:
        type: integer
      value:
        type: string
      description:
        type: string
      descriptionOfAboveFourOption:
         type: string
      scoreType:
         type: array
         items:
          $ref: '#/definitions/ScoreTypeDTO'
  SectionDTO:
   type: object
   properties:
      id:
        type: integer
      scoreType:
        $ref: '#/definitions/ScoreTypeDTO'
      name:
        type: string
      description:
        type: string
      position:
        type: integer
      questions:
        type: array
        items:
          $ref: '#/definitions/QuestionDTO'
...